oidc_backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
gunicorn==21.2.0
psycopg2-binary==2.9.9
alembic==1.13.1
PyJWT==2.8.0
cryptography==41.0.7
cbor2==5.4.7
pycose==1.0.0
pymdoccbor==0.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1 